fix(ui): keep automation deep links off startup
