+++ src/config/control-ui-link-base.ts
@@ -1,9 +1,9 @@
 import {
-  buildControlUiAutomationPath,
   buildControlUiSessionPath,
   normalizeControlUiBasePath,
 } from "@openclaw/session-url-contract";
-import type { ControlUiAutomationTab } from "@openclaw/session-url-contract";
+import { buildControlUiAutomationPath } from "@openclaw/session-url-contract/automation.runtime";
+import type { ControlUiAutomationTab } from "@openclaw/session-url-contract/automation.runtime";
 import { resolveGatewayPublicOrigin } from "./gateway-public-origin.js";
 import type { OpenClawConfig } from "./types.js";
