+++ ui/src/pages/cron/route.ts
@@ -1,23 +1,18 @@
-import { definePage, type RouteLocation } from "@openclaw/uirouter";
-import { html } from "lit";
-import { routePageSpec } from "../../app-route-paths.ts";
+import { definePage } from "@openclaw/uirouter";
+import { INTERNAL_AUTOMATION_PATH_PARAM, routePageSpec } from "../../app-route-paths.ts";
 import type { ApplicationContext } from "../../app/context.ts";
-import { cronRouteLocation, resolveCronRouteData, type CronRouteData } from "./route-location.ts";
 
-export type { CronRouteData } from "./route-location.ts";
+export type { CronRouteData } from "./route-location.runtime.ts";
+
+const loadCronPage = () => import("./cron-page.ts");
 
 export const page = definePage({
   ...routePageSpec("cron"),
-  loaderDeps: (_context: ApplicationContext, location: RouteLocation) => {
-    const routeLocation = cronRouteLocation(location);
-    return `${routeLocation.pathname}\u0000${routeLocation.search}\u0000${routeLocation.hash}`;
-  },
-  loader: (context: ApplicationContext, { location }) =>
-    resolveCronRouteData(location, context.basePath),
-  component: () =>
-    import("./cron-page.ts").then(() => ({
-      header: true,
-      render: (data: CronRouteData | undefined) =>
-        html`<openclaw-cron-page .routeData=${data}></openclaw-cron-page>`,
-    })),
+  loaderDeps: (_context: ApplicationContext, location) =>
+    new URLSearchParams(location.search).get(INTERNAL_AUTOMATION_PATH_PARAM) ?? location.pathname,
+  // The URL grammar is automation-only code; load it with the page data rather
+  // than pulling the deep-link helper into every Control UI startup.
+  loader: (context: ApplicationContext, options) =>
+    loadCronPage().then(({ loadCronRouteData }) => loadCronRouteData(context, options)),
+  component: loadCronPage,
 });
