+++ ui/src/test-helpers/control-ui-e2e.ts
@@ -77,6 +77,21 @@ type ControlUiRouteTarget = {
   search?: string;
 };
 
+export function controlUiPathnamesEqual(left: string, right: string): boolean {
+  const leftSegments = left.split("/");
+  const rightSegments = right.split("/");
+  if (leftSegments.length !== rightSegments.length) {
+    return false;
+  }
+  try {
+    return leftSegments.every(
+      (segment, index) => decodeURIComponent(segment) === decodeURIComponent(rightSegments[index]!),
+    );
+  } catch {
+    return false;
+  }
+}
+
 // Cold Vite route chunks can monopolize Chromium on loaded CI hosts. Keep the
 // wait browser-local, but allow enough time for the router to finish committing.
 const CONTROL_UI_ROUTE_TIMEOUT_MS = 60_000;
@@ -95,6 +110,21 @@ export async function waitForControlUiRoute(page: Page, target: ControlUiRouteTa
   try {
     const handle = await page.waitForFunction(
       (expected) => {
+        const pathnamesEqual = (left: string, right: string) => {
+          const leftSegments = left.split("/");
+          const rightSegments = right.split("/");
+          if (leftSegments.length !== rightSegments.length) {
+            return false;
+          }
+          try {
+            return leftSegments.every(
+              (segment, index) =>
+                decodeURIComponent(segment) === decodeURIComponent(rightSegments[index]!),
+            );
+          } catch {
+            return false;
+          }
+        };
         const app = document.querySelector("openclaw-app") as HTMLElement & {
           runtime?: {
             router: {
@@ -112,9 +142,10 @@ export async function waitForControlUiRoute(page: Page, target: ControlUiRouteTa
         return (
           state?.status === "success" &&
           state.matches[0]?.routeId === expected.routeId &&
-          state.resolvedLocation?.pathname === pathname &&
+          state.resolvedLocation !== null &&
+          pathnamesEqual(state.resolvedLocation.pathname, pathname) &&
           state.pendingMatches.length === 0 &&
-          (expected.pathname === undefined || pathname === expected.pathname) &&
+          (expected.pathname === undefined || pathnamesEqual(pathname, expected.pathname)) &&
           (expected.pathnamePrefix === undefined || pathname.startsWith(expected.pathnamePrefix)) &&
           (expected.search === undefined || window.location.search === expected.search) &&
           (expected.hash === undefined || window.location.hash === expected.hash)
